requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-cors-headers==4.3.1
scikit-learn==1.3.2
shap==0.43.0
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
redis==5.0.1
python-dotenv==1.0.0
whitenoise==6.6.0
